+++ angel-ps/core/src/main/java/com/tencent/angel/worker/task/TaskManager.java
@@ -214,7 +214,7 @@ public class TaskManager {
       int lastPos = 0;
       for (int i = 0; i < partNum; i++) {
         PartitionKey partKey = partKeyList.get(i);
-        int endCol = partKey.getEndCol();
+        long endCol = partKey.getEndCol();
         for (int j = lastPos; j < size; j++) {
           if (indexArray[j] >= endCol) {
